refatoracao de codigo | atualizar autoescola | ajuste uploadArquivo

+++ NET/AulaRemota.Core/Arquivo/Deletar/ArquivoDeletarHandler.cs
@@ -3,30 +3,45 @@ using Microsoft.WindowsAzure.Storage.Blob;
 using Microsoft.WindowsAzure.Storage;
 using System.Threading;
 using System.Threading.Tasks;
+using AulaRemota.Shared.Helpers.Constants;
+using AulaRemota.Shared.Helpers;
 
 namespace AulaRemota.Core.Arquivo.Deletar
 {
-    class ArquivoDeletarHandler : IRequestHandler<ArquivoDeletarInput,bool>
+    public class ArquivoDeletarHandler : IRequestHandler<ArquivoDeletarInput, bool>
     {
         CloudStorageAccount cloudStorageAccount = CloudStorageAccount.Parse("DefaultEndpointsProtocol=https;AccountName=aularemotablob;AccountKey=OBPERY2qeFupdEalWpLTXXlNmEWvQVKljJWkYgrr1dTOalbL/R/KxF82qEdk+4rVD2mx2CR28KqS8vkxvn5jkg==;EndpointSuffix=core.windows.net");
         async Task<bool> IRequestHandler<ArquivoDeletarInput, bool>.Handle(ArquivoDeletarInput request, CancellationToken cancellationToken)
         {
-            var cloudBlobClient = cloudStorageAccount.CreateCloudBlobClient();
-            var cloudBlobContainer = cloudBlobClient.GetContainerReference("files");
-            if (request.NivelAcesso >= 20 && request.NivelAcesso < 30) cloudBlobContainer = cloudBlobClient.GetContainerReference("autoescola");
-
-            if (await cloudBlobContainer.CreateIfNotExistsAsync())
+            if (request.Arquivos.Count > 0)
             {
-                await cloudBlobContainer.SetPermissionsAsync(new BlobContainerPermissions
+                var cloudBlobClient = cloudStorageAccount.CreateCloudBlobClient();
+                var cloudBlobContainer = cloudBlobClient.GetContainerReference("files");
+                if (request.TipoUsuario.Equals(Constants.Roles.AUTOESCOLA)) cloudBlobContainer = cloudBlobClient.GetContainerReference(nameof(Constants.Roles.AUTOESCOLA).ToLower());
+                if (request.TipoUsuario.Equals(Constants.Roles.ADMINISTRATIVO)) cloudBlobContainer = cloudBlobClient.GetContainerReference(nameof(Constants.Roles.ADMINISTRATIVO).ToLower());
+                if (request.TipoUsuario.Equals(Constants.Roles.ALUNO)) cloudBlobContainer = cloudBlobClient.GetContainerReference(nameof(Constants.Roles.ALUNO).ToLower());
+                if (request.TipoUsuario.Equals(Constants.Roles.EDRIVING)) cloudBlobContainer = cloudBlobClient.GetContainerReference(nameof(Constants.Roles.EDRIVING).ToLower());
+                if (request.TipoUsuario.Equals(Constants.Roles.INSTRUTOR)) cloudBlobContainer = cloudBlobClient.GetContainerReference(nameof(Constants.Roles.INSTRUTOR).ToLower());
+                if (request.TipoUsuario.Equals(Constants.Roles.PARCEIRO)) cloudBlobContainer = cloudBlobClient.GetContainerReference(nameof(Constants.Roles.PARCEIRO).ToLower());
+
+                if (await cloudBlobContainer.CreateIfNotExistsAsync())
                 {
-                    PublicAccess = BlobContainerPublicAccessType.Off
-                });
+                    await cloudBlobContainer.SetPermissionsAsync(new BlobContainerPermissions
+                    {
+                        PublicAccess = BlobContainerPublicAccessType.Off
+                    });
+                }
+
+                foreach (var item in request.Arquivos)
+                {
+                    CloudBlockBlob blob = cloudBlobContainer.GetBlockBlobReference(item.Nome);
+
+                    var result = await blob.DeleteIfExistsAsync();
+                    if (!result)
+                        throw new CustomException(item.Nome + "não pode ser removido");
+                }
             }
-            CloudBlockBlob blob = cloudBlobContainer.GetBlockBlobReference(request.NomeArquivo);
-
-            var result = await blob.DeleteIfExistsAsync();
-
-            return result;
+            return true;
         }
     }
 }
